microservices/course_ingestion/requirements.txt
tornado==6.3.2
jsonschema==2.6.0
firebase-admin==4.5.2
Scrapy==2.8.0
selenium==3.141.0
html2text==2020.1.16
webdriver-manager==3.3.0
fireo==1.4.1
spacy==3.2.0
PyMuPDF==1.18.14
pytesseract==0.3.8
umap-learn==0.5.2
scikit-learn==0.24.2
sentencepiece==0.1.96
sentence-transformers==2.2.2
textacy==0.11.0
pyyaml==6.0
kubernetes==11.0.0
absl-py==0.11.0
py2neo==2021.2.0
protobuf==3.19.5
google-cloud-logging==3.1.2
redis==4.5.4
gcsfs==2022.5.0
editdistance==0.6.0
Twisted==22.10.0
pyOpenSSL==23.0.0
huggingface-hub==0.10.0
networkx==2.8.8
typing-extensions==4.0.1